need help with transfer

+++ contracts/LBP/LBPWrapper.sol
@@ -89,7 +89,7 @@ contract LBPWrapper {
         // to handle stack overflow
         {
             address vault = address(ILBPFactory(LBPFactory).getVault());
-            for ( uint i; i < _tokens.length; i++ ) {
+            for ( uint8 i; i < _tokens.length; i++ ) {
                 IERC20(_tokens[i]).approve(vault, _amounts[i]);
             }
         }
