+++ contracts/LBP/WrapperFactory.sol
@@ -93,7 +93,7 @@ contract WrapperFactory is CloneFactory, Ownable {
 
         LBPWrapper(wrapper).initialize(LBPFactory, swapFeePercentage);
 
-        for (uint i; i < _tokens.length; i++) {
+        for (uint8 i; i < _tokens.length; i++) {
             IERC20(_tokens[i]).transferFrom(_admin, wrapper, _amounts[i]);
         }
